Refactor for clippy
