+++ token-swap/program/src/entrypoint.rs
@@ -3,7 +3,7 @@
 #![cfg(feature = "program")]
 #![cfg(not(feature = "no-entrypoint"))]
 
-use crate::{error::Error, state::State};
+use crate::{error::Error, processor::Processor};
 use solana_sdk::{
     account_info::AccountInfo, entrypoint, entrypoint::ProgramResult,
     program_error::PrintProgramError, pubkey::Pubkey,
@@ -15,7 +15,7 @@ fn process_instruction<'a>(
     accounts: &'a [AccountInfo<'a>],
     instruction_data: &[u8],
 ) -> ProgramResult {
-    if let Err(error) = State::process(program_id, accounts, instruction_data) {
+    if let Err(error) = Processor::process(program_id, accounts, instruction_data) {
         // catch the error so we can print it
         error.print::<Error>();
         return Err(error);
